requirements.txt
altair==5.3.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
six==1.16.0
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.1